Replace fp16 with bf16 in training configuration for improved compatibility

+++ src/train.py
@@ -143,7 +143,7 @@ def main():
         metric_for_best_model="eval_loss",
         greater_is_better=False,
         predict_with_generate=False,
-        fp16=torch.cuda.is_available() and args.method != "qlora",
+        bf16=torch.cuda.is_available() and args.method != "qlora",
         fp16=False,
         report_to="none",
     )
